Needle in Haystack
This program searches an array for a string == 'needle' and returns the message 'found the needle at position (the needle's position"

I suspect it would be more valuable and reusable as a function if it searched for any string rather than just 'needle'--when i revisit the file I'll probably update it to incorporate this idea. For now, solving the declared puzzle was more interesting.